[CSL-1621] Split rubbish command types and parser into two modules
